Use extension specific interfaces for DI
