+++ src/Azure.Deployments.Extensibility.Extensions.Kubernetes/DependencyInjection/IServiceCollectionExtensions.cs
@@ -2,8 +2,6 @@
 // Licensed under the MIT License.
 
 using Azure.Deployments.Extensibility.AspNetCore;
-using Azure.Deployments.Extensibility.Core.V2.Models;
-using Azure.Deployments.Extensibility.Core.V2.Validation;
 using Azure.Deployments.Extensibility.Extensions.Kubernetes.Client;
 using Azure.Deployments.Extensibility.Extensions.Kubernetes.Validation;
 using Microsoft.Extensions.DependencyInjection;
@@ -13,8 +11,8 @@ namespace Azure.Deployments.Extensibility.Extensions.Kubernetes.DependencyInject
     public static class IServiceCollectionExtensions
     {
         public static IServiceCollection AddKubernetesExtensionDispatcher(this IServiceCollection services) => services
-            .AddSingleton<IModelValidator<ResourceSpecification>, ResourceSpecificationValidator>()
-            .AddSingleton<IModelValidator<ResourceReference>, ResourceReferenceValidator>()
+            .AddSingleton<IK8sResourceSpecificationValidator, K8sResourceSpecificationValidator>()
+            .AddSingleton<IK8sResourceReferenceValidator, K8sResourceReferenceValidator>()
             .AddSingleton<IK8sClientFactory, K8sClientFactory>()
             .AddKeyedSingleton<IExtensionDispatcher, KubernetesExtensionDispatcher>(KubernetesExtension.ExtensionName);
     }
