+++ src/Azure.Deployments.Extensibility.Extensions.Kubernetes/KubernetesExtensionDispatcher.cs
@@ -3,22 +3,21 @@
 
 using Azure.Deployments.Extensibility.AspNetCore;
 using Azure.Deployments.Extensibility.AspNetCore.Exceptions;
-using Azure.Deployments.Extensibility.Core.V2.Models;
-using Azure.Deployments.Extensibility.Core.V2.Validation;
 using Azure.Deployments.Extensibility.Extensions.Kubernetes.Client;
+using Azure.Deployments.Extensibility.Extensions.Kubernetes.Validation;
 using Semver;
 
 namespace Azure.Deployments.Extensibility.Extensions.Kubernetes
 {
     public class KubernetesExtensionDispatcher : IExtensionDispatcher
     {
-        private readonly IModelValidator<ResourceSpecification> resourceSpecificationValidator;
-        private readonly IModelValidator<ResourceReference> resourceReferenceValidator;
+        private readonly IK8sResourceSpecificationValidator resourceSpecificationValidator;
+        private readonly IK8sResourceReferenceValidator resourceReferenceValidator;
         private readonly IK8sClientFactory clientFactory;
 
         public KubernetesExtensionDispatcher(
-            IModelValidator<ResourceSpecification> resourceSpecificationValidator,
-            IModelValidator<ResourceReference> resourceReferenceValidator,
+            IK8sResourceSpecificationValidator resourceSpecificationValidator,
+            IK8sResourceReferenceValidator resourceReferenceValidator,
             IK8sClientFactory clientFactory)
         {
             this.resourceSpecificationValidator = resourceSpecificationValidator;
