+++ src/Azure.Deployments.Extensibility.Extensions.Kubernetes/Validation/IK8sResourceSpecificationValidator.cs
@@ -1,0 +1,12 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+using Azure.Deployments.Extensibility.Core.V2.Models;
+using Azure.Deployments.Extensibility.Core.V2.Validation;
+
+namespace Azure.Deployments.Extensibility.Extensions.Kubernetes.Validation
+{
+    public interface IK8sResourceSpecificationValidator : IModelValidator<ResourceSpecification>
+    {
+    }
+}
